snipping agent

+++ OnlineAuctionSystem/OnlineAuctionSystem-ejb/src/java/ejb/session/stateless/AuctionEntityControllerLocal.java
@@ -44,7 +44,7 @@ public interface AuctionEntityControllerLocal {
 
     public BidEntity getCurrentWinningBidEntity(Long productid) throws AuctionNotFoundException;
 
-    public Double getCurrentBidIncremental(BigDecimal currentprice);
+    public BigDecimal getCurrentBidIncremental(BigDecimal currentprice);
 
     public BidEntity placeNewBid(Long productid, CustomerEntity customer) throws AuctionNotFoundException, BidAlreadyExistException, GeneralException;
